Move the tool window control to the Views folder.
