+++ src/Sarif.Viewer.VisualStudio/Sarif.Viewer.VisualStudio.csproj
@@ -73,7 +73,7 @@
     </Compile>
     <Compile Include="SarifToolWindow.cs" />
     <Compile Include="SarifToolWindowCommand.cs" />
-    <Compile Include="SarifToolWindowControl.xaml.cs">
+    <Compile Include="Views\SarifToolWindowControl.xaml.cs">
       <DependentUpon>SarifToolWindowControl.xaml</DependentUpon>
     </Compile>
     <Compile Include="Sarif\Replacement.extensions.cs" />
@@ -323,7 +323,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="SarifToolWindowControl.xaml">
+    <Page Include="Views\SarifToolWindowControl.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
