+++ src/Sarif.Viewer.VisualStudio/Views/SarifToolWindowControl.xaml
@@ -1,15 +1,15 @@
-﻿<UserControl x:Class="Microsoft.Sarif.Viewer.SarifToolWindowControl"
+﻿<UserControl x:Class="Microsoft.Sarif.Viewer.Views.SarifToolWindowControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:views="clr-namespace:Microsoft.Sarif.Viewer.Views"
+             xmlns:local="clr-namespace:Microsoft.Sarif.Viewer.Views"
              Background="{DynamicResource VsBrush.Window}"
              Foreground="{DynamicResource VsBrush.WindowText}"
              mc:Ignorable="d"
              d:DesignHeight="300" d:DesignWidth="300"
              Name="MyToolWindow">
     <Grid>
-        <views:CodeLocations Visibility="{Binding TargetNullValue=Collapsed}" />
+        <local:CodeLocations Visibility="{Binding TargetNullValue=Collapsed}" />
     </Grid>
 </UserControl>
